replace gulp with webpack assets bundling
