--- demo/css/base.css
@@ -1,1 +1,0 @@
-article,aside,details,figcaption,figure,footer,header,hgroup,menu,nav,pre,section,small,ul{display:block}a,abbr,acronym,address,applet,article,aside,audio,b,big,blockquote,body,canvas,caption,center,cite,code,dd,del,details,dfn,div,dl,dt,em,embed,fieldset,figcaption,figure,footer,form,h1,h2,h3,h4,h5,h6,header,hgroup,html,i,iframe,img,ins,kbd,label,legend,li,mark,menu,nav,object,ol,output,p,pre,q,ruby,s,samp,section,small,span,strike,strong,sub,summary,sup,table,tbody,td,tfoot,th,thead,time,tr,tt,u,ul,var,video{margin:0;padding:0;border:0;font:inherit;font-size:100%;vertical-align:baseline}.main-label,h1,h2,h3,label.current,strong{font-weight:600}ul{list-style:disc inside;margin-left:20px;margin-bottom:20px}blockquote,q{quotes:none}blockquote:after,blockquote:before,q:after,q:before{content:'';content:none}table{border-collapse:collapse;border-spacing:0}html{box-sizing:border-box;font-size:62.5%}*,:after,:before{box-sizing:inherit}body{font-size:1.6rem;line-height:1.4;font-family:'Source Sans Pro',sans-serif;color:#323232;background:#f5f5f5}a{color:inherit}small{font-size:1rem;text-align:right}h1{font-size:4rem;margin-bottom:20px}h2,p.about{font-size:2.5rem;margin-bottom:20px}p{margin-bottom:15px}h3{font-size:2rem;margin-bottom:20px}.uo{color:#b4b4b4}code,pre{background-color:#dcdcdc;font-family:'Source Code Pro',monospace;color:#000}ul li{margin-bottom:10px}.container{margin:100px auto;width:90%;max-width:800px}@media screen and (max-width:600px){.container{margin:50px auto}}section.site-section{margin-bottom:50px;padding-bottom:30px;border-bottom:1px solid #dcdcdc}section:not(.site-section){margin-bottom:40px}.highlight{margin-bottom:15px}pre{padding:10px;font-size:1.2rem;white-space:pre-line;word-wrap:break-word;position:relative}code{font-size:1.4rem;padding:0 5px}.copy-code{display:inline-block;width:auto;border:2px solid #dcdcdc;border-top:0;background-color:#fafafa;padding:0 10px;cursor:pointer;float:right}.copy-code:hover{background-color:#dcdcdc}.input-group{margin-bottom:20px}label{display:block}.main-label{margin-bottom:5px}label:not(.main-label){padding:5px 0}input[type=submit],select{height:40px;display:inline-block;font-size:1.6rem;font-family:'Source Sans Pro',sans-serif}select{padding:5px;background-color:#fff;border:2px solid #DB5600;margin-right:10px;margin-bottom:10px;width:100%;max-width:400px}input[type=submit]{width:auto;cursor:pointer;padding:5px 10px;background-color:#DB5600;border:0;border-radius:5px;color:#fff}.twa{display:inline-block;height:1em;width:1em;margin:0 .05em 0 .1em;vertical-align:-.1em;background-repeat:no-repeat;background-position:center center;background-size:1em 1em}.twa-heart{background-image:url(https://twemoji.maxcdn.com/svg/2764.svg)}button[id^=trigger-example]{margin-bottom:15px}
