+++ src/css/base.css
@@ -1,0 +1,263 @@
+/* RESET */
+html, body, div, span, applet, object, iframe,
+h1, h2, h3, h4, h5, h6, p, blockquote, pre,
+a, abbr, acronym, address, big, cite, code,
+del, dfn, em, img, ins, kbd, q, s, samp,
+small, strike, strong, sub, sup, tt, var,
+b, u, i, center,
+dl, dt, dd, ol, ul, li,
+fieldset, form, label, legend,
+table, caption, tbody, tfoot, thead, tr, th, td,
+article, aside, canvas, details, embed,
+figure, figcaption, footer, header, hgroup,
+menu, nav, output, ruby, section, summary,
+time, mark, audio, video {
+	margin: 0;
+	padding: 0;
+	border: 0;
+	font: inherit;
+	font-size: 100%;
+	vertical-align: baseline;
+}
+article, aside, details, figcaption, figure,
+footer, header, hgroup, menu, nav, section {
+	display: block;
+}
+ul {
+	list-style: none;
+}
+blockquote, q {
+	quotes: none;
+}
+blockquote:before, blockquote:after,
+q:before, q:after {
+	content: '';
+	content: none;
+}
+table {
+	border-collapse: collapse;
+	border-spacing: 0;
+}
+html {
+  box-sizing: border-box;
+}
+*, *:before, *:after {
+  box-sizing: inherit;
+}
+
+
+
+/* ---------------*/
+html {
+    font-size: 62.5%;
+}
+body {
+    font-size: 1.6rem;
+    line-height: 1.4;
+    font-family: 'Source Sans Pro', sans-serif;
+    color: rgb(50, 50, 50);
+    /*background: rgb(245,245,245);*/
+
+}
+a {
+    color: inherit; 
+}
+
+small {
+	font-size: 1rem;
+	display: block;
+	text-align: right;
+}
+strong {
+	font-weight: 600;
+}
+
+h1 {
+	font-size: 4rem;
+	margin-bottom: 20px;
+	font-weight: 600;
+}
+
+
+
+
+p.about {
+	font-size: 2.5rem;
+	margin-bottom: 20px;
+}
+
+p {
+	margin-bottom: 15px;
+}
+
+h2 {
+	font-size: 2.5rem;
+	font-weight: 600;
+	margin-bottom: 20px;
+}
+
+
+h3 {
+	font-size: 2rem;
+	font-weight: 600;
+	margin-bottom: 20px;
+}
+
+.uo {
+	color: rgb(180, 180, 180);
+}
+
+ul {
+	list-style-type: disc;
+	list-style-position: inside;
+	margin-left: 20px;
+	display: block;
+	margin-bottom: 20px;
+}
+
+ul li {
+	margin-bottom: 10px;
+}
+
+
+.container {
+	margin: 100px auto;
+	width: 90%;
+	max-width: 800px;
+}
+
+@media screen and (max-width: 600px) {
+	.container {
+		margin: 50px auto;
+	}
+}
+
+
+
+section.site-section {
+	margin-bottom: 50px;
+	padding-bottom: 30px;
+	border-bottom: 1px solid rgb(220, 220, 220);
+}
+section:not(.site-section) {
+	margin-bottom: 40px;
+}
+
+.highlight {
+	margin-bottom: 15px;
+}
+
+pre {
+	background-color: rgb(220, 220, 220);
+	padding: 10px;
+	display: block;
+	font-family: 'Source Code Pro', monospace;
+	font-size: 1.2rem;
+	color: #000;
+	
+	white-space: pre-line;
+	word-wrap: break-word;
+	position: relative;
+}
+
+
+code {
+	background-color: rgb(220, 220, 220);
+	font-family: 'Source Code Pro', monospace;
+	font-size: 1.4rem;
+	color: #000;
+	padding: 0 5px;
+}
+
+
+
+.copy-code {
+	display: inline-block;
+	width: auto;
+	border: 2px solid rgb(220, 220, 220);
+	border-top: 0;
+	background-color: rgb(250, 250, 250);
+	padding: 0px 10px;
+	cursor: pointer;
+	float: right;
+}
+.copy-code:hover {
+	background-color: rgb(220, 220, 220);
+}
+
+.input-group {
+	margin-bottom: 20px;
+}
+
+label {
+	display: block;
+}
+
+.main-label {
+	font-weight: 600;
+	margin-bottom: 5px;
+}
+
+label:not(.main-label) {
+	padding: 5px 0;
+}
+
+select,
+input[type="submit"] {
+	height: 40px;	
+	display: inline-block;
+	font-size: 1.6rem;
+    font-family: 'Source Sans Pro', sans-serif;
+}
+
+select {
+	padding: 5px;
+	background-color: #fff;
+	border: 2px solid #DB5600;
+	margin-right: 10px;
+	margin-bottom: 10px;
+
+	width: 100%;
+	max-width: 400px;
+
+}
+
+input[type="submit"] {
+	width: auto;
+	cursor: pointer;
+	padding: 5px 10px;
+	background-color: #DB5600;
+	border: 0;
+	border-radius: 5px;
+	color: #fff;
+
+}
+
+label.current {
+	font-weight: 600;
+}
+
+input[type="checkbox"] {
+}
+
+
+/* Twemoji Awesome - http://ellekasai.github.io/twemoji-awesome/ */
+ .twa {
+  display: inline-block;
+  height: 1em;
+  width: 1em;
+  margin: 0 .05em 0 .1em;
+  vertical-align: -0.1em;
+  background-repeat: no-repeat;
+  background-position: center center;
+  background-size: 1em 1em; }
+
+.twa-heart {
+  background-image: url("https://twemoji.maxcdn.com/svg/2764.svg"); 
+}
+
+
+button[id^="trigger-example"] {
+	margin-bottom: 15px;
+}
+
